- added help text to CommandLineWrapper class to show usage when user passed wrong arguments - refactored classes paths and names to uphold onion arch and SOLID
