+++ src/BluePrism.WordLadder.Infrastructure/CommandLineHelpers/CommandLineWrapper.cs
@@ -6,39 +6,35 @@ namespace BluePrism.WordLadder.Infrastructure.CommandLineHelpers
 {
     public class CommandLineWrapper : ICommandLineWrapper
     {
-        private readonly ParserResult<Options> _resultParsing;
+        private readonly Parser _parser;
+        private readonly string[] _args;
+        private ParserResult<Options> _resultParsing;
         private Options _resultArgs;
 
         public CommandLineWrapper(string[] args)
         {
-            // (1) default options
-            _resultParsing = Parser.Default.ParseArguments<Options>(args);
-        }
-
-        public void ValidateResult()
-        {
-            _resultParsing
-                .WithParsed(HandleOptions)
-                .WithNotParsed(HandleErrors);
-        }
-
-        private void HandleOptions(Options opt)
-        {
-            Console.WriteLine(opt.BeginWord);
-        }
-
-        private void HandleErrors(IEnumerable<Error> obj)
-        {
-            Console.WriteLine(obj.Output());
+            _parser = new Parser(config => config.HelpWriter = Console.Out);
+            _args = args;
         }
 
         public Options GetResult()
         {
-            _resultParsing.WithParsed(opt =>
-            {
-                _resultArgs = opt;
-            });
+            _resultParsing = _parser.ParseArguments<Options>(_args);
+            _resultParsing
+                .WithParsed(HandleOptions)
+                .WithNotParsed(HandleErrors);
+
             return _resultArgs;
+        }
+
+        private void HandleOptions(Options opt)
+        {
+            _resultArgs = opt;
+        }
+
+        private void HandleErrors(IEnumerable<Error> obj)
+        {
+            obj.Output();
         }
     }
 }
