+++ src/BluePrism.WordLadder.Infrastructure/CommandLineHelpers/ICommandLineWrapper.cs
@@ -2,7 +2,6 @@
 {
     public interface ICommandLineWrapper
     {
-        void ValidateResult();
         Options GetResult();
     }
 }
