+++ src/BluePrism.WordLadder.Test/BluePrism.WordLadder.Test.csproj
@@ -19,6 +19,8 @@
     <PackageReference Include="FluentAssertions" Version="5.10.3" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\BluePrism.WordLadder.Domain\BluePrism.WordLadder.Domain.csproj" />
+    <ProjectReference Include="..\BluePrism.WordLadder.Infrastructure\BluePrism.WordLadder.Infrastructure.csproj" />
     <ProjectReference Include="..\BluePrism.WordLadder\BluePrism.WordLadder.csproj" />
   </ItemGroup>
   <ItemGroup>
